pdf_service/requirements.txt
fastapi
uvicorn
reportlab
pillow
psycopg2-binary
python-dotenv
pydantic
